+++ classes/ListFilterForm.php
@@ -125,5 +125,15 @@ class ListFilterForm extends Form
         }
 
         $this->objListModule->modifyDC($arrDca);
+	
+		// HOOK: add custom logic
+		if (isset($GLOBALS['TL_HOOKS']['modifyDCFilter']) && is_array($GLOBALS['TL_HOOKS']['modifyDCFilter']))
+		{
+			foreach ($GLOBALS['TL_HOOKS']['modifyDCFilter'] as $callback)
+			{
+				$this->import($callback[0]);
+				$this->{$callback[0]}->{$callback[1]}($arrDca);
+			}
+		}
     }
 }