+++ languages/de/modules.php
@@ -9,3 +9,4 @@ $GLOBALS['TL_LANG']['FMD'][MODULE_FORMHYBRID_LISTS] = ['FormHybrid-Listen', ''];
 $GLOBALS['TL_LANG']['FMD'][MODULE_FORMHYBRID_LIST] = ['FormHybrid-Liste', ''];
 $GLOBALS['TL_LANG']['FMD'][MODULE_FORMHYBRID_MEMBER_LIST] = ['FormHybrid-Mitgliederliste', ''];
 $GLOBALS['TL_LANG']['FMD'][MODULE_FORMHYBRID_NEWS_LIST] = ['FormHybrid-Nachrichtenliste', ''];
+$GLOBALS['TL_LANG']['FMD'][MODULE_FORMHYBRID_LIST_FILTER] = ['FormHybrid-Filter', ''];