+++ models/FormHybridFilterModel.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace HeimrichHannot\FormHybridList;
+
+abstract class FormHybridFilterModel extends \Model
+{
+	protected static $strTable = 'tl_module';
+}